Add files via upload
Project : You’ll implement a simple machine learning algorithm in Python using Scikit-learn, a machine learning tool for Python. Using a database of breast cancer tumor information, you’ll use a Naive Bayes (NB) classifier that predicts whether or not a tumor is malignant or benign.
Data: Scikit-learn comes installed with various datasets which we can load into Python, and the dataset we want is included.